Move SanitizeDates concern
This moves the concern out of the TeacherInterface and in to the global
scope so we can use it in multiple places.
